+++ runhaz.sh
@@ -110,12 +110,10 @@ graph_models() {
     print_style "Generating Entity Relationship Diagram...\n" "success";
     exec_cmd="$base_py_cmd graph_models sites trak \
     -g \
-    --pygraphviz \
     --settings haztrak.settings \
-    --hide-edge-labels \
     --rankdir=RL \
     --arrow-shape=normal \
-    -E \
+    --exclude-models=*BaseModel \
     --disable-abstract-fields \
     -o $base_dir/docs/haztrak_book/src/assets/erd.png"
     eval "$exec_cmd"
