+++ server/apps/sites/views/site_views.py
@@ -10,8 +10,9 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 
 from apps.sites.models import Site
+from apps.sites.serializers import SiteSerializer
 from apps.trak.models import Manifest
-from apps.trak.serializers import MtnSerializer, SiteSerializer
+from apps.trak.serializers import MtnSerializer
 from apps.trak.tasks import sync_site_manifests
 
 logger = logging.getLogger(__name__)
