+++ server/apps/trak/models/__init__.py
@@ -5,7 +5,7 @@ Model definitions for the trak domain
 from .code_models import WasteCode
 from .contact_models import EpaPhone
 from .handler_models import ManifestHandler
-from .manifest_models import Manifest
+from .manifest_models import AdditionalInfo, Manifest
 from .signature_models import ESignature, PaperSignature, QuickerSign, Signer
 from .transporter_models import Transporter
 from .waste_models import WasteLine
