+++ server/apps/trak/serializers/__init__.py
@@ -1,13 +1,4 @@
-from .address_ser import AddressSerializer
-from .contact_ser import ContactSerializer, EpaPhoneSerializer
-from .handler_ser import EpaSiteSerializer, ManifestHandlerSerializer
+from .handler_ser import ManifestHandlerSerializer, TransporterSerializer
 from .manifest_ser import ManifestSerializer, MtnSerializer
-from .rcra_profile_ser import (
-    EpaPermissionSerializer,
-    ProfileGetSerializer,
-    ProfileUpdateSerializer,
-    SitePermissionSerializer,
-)
-from .sites_ser import SiteSerializer
-from .transporter_ser import TransporterSerializer
+from .signature_ser import ESignatureSerializer, PaperSignatureSerializer, QuickerSignSerializer
 from .waste_line_ser import WasteCodeSerializer, WasteLineSerializer
