+++ server/apps/trak/serializers/handler_ser.py
@@ -1,95 +1,12 @@
 from typing import Dict
 
 from rest_framework import serializers
+from rest_framework.exceptions import ValidationError
 
-from apps.sites.models import EpaSite
-from apps.trak.models import ManifestHandler
-from apps.trak.serializers import AddressSerializer
+from apps.sites.serializers import EpaSiteSerializer
+from apps.trak.models import ManifestHandler, Transporter
 
-from .base_ser import TrakBaseSerializer
-from .contact_ser import ContactSerializer, EpaPhoneSerializer
 from .signature_ser import ESignatureSerializer, PaperSignatureSerializer
-
-
-class EpaSiteSerializer(TrakBaseSerializer):
-    """
-    EpaSite model serializer for JSON marshalling/unmarshalling
-    """
-
-    epaSiteId = serializers.CharField(
-        source="epa_id",
-    )
-    siteType = serializers.CharField(
-        source="site_type",
-        allow_null=True,
-        required=False,
-    )
-    modified = serializers.BooleanField(
-        allow_null=True,
-        default=False,
-    )
-    # name
-    mailingAddress = AddressSerializer(
-        source="mail_address",
-    )
-    siteAddress = AddressSerializer(
-        source="site_address",
-    )
-    contact = ContactSerializer()
-    emergencyPhone = EpaPhoneSerializer(
-        source="emergency_phone",
-        allow_null=True,
-        default=None,
-    )
-    # paperSignatureInfo
-    registered = serializers.BooleanField(
-        allow_null=True,
-        default=False,
-    )
-    limitedEsign = serializers.BooleanField(
-        source="limited_esign",
-        allow_null=True,
-        default=False,
-    )
-    canEsign = serializers.BooleanField(
-        source="can_esign",
-        allow_null=True,
-        default=False,
-    )
-    hasRegisteredEmanifestUser = serializers.BooleanField(
-        source="registered_emanifest_user",
-        allow_null=True,
-        default=False,
-    )
-    gisPrimary = serializers.BooleanField(
-        source="gis_primary",
-        allow_null=True,
-        default=False,
-    )
-
-    def update(self, instance, validated_data):
-        return self.Meta.model.objects.save(**validated_data)
-
-    def create(self, validated_data):
-        return self.Meta.model.objects.save(**validated_data)
-
-    class Meta:
-        model = EpaSite
-        fields = [
-            "epaSiteId",
-            "siteType",
-            "modified",
-            "name",
-            "siteAddress",
-            "mailingAddress",
-            "contact",
-            "emergencyPhone",
-            "registered",
-            "limitedEsign",
-            "canEsign",
-            "hasRegisteredEmanifestUser",
-            "gisPrimary",
-        ]
 
 
 class ManifestHandlerSerializer(EpaSiteSerializer):
@@ -139,3 +56,27 @@ class ManifestHandlerSerializer(EpaSiteSerializer):
             "paperSignatureInfo",
             "signed",
         ]
+
+
+class TransporterSerializer(ManifestHandlerSerializer):
+    """
+    Transporter model serializer for JSON marshalling/unmarshalling
+    """
+
+    class Meta:
+        model = Transporter
+        fields = [
+            "epa_site",
+            "order",
+            "paperSignatureInfo",
+            "electronicSignaturesInfo",
+            "signed",
+        ]
+
+    def to_internal_value(self, data):
+        """Move fields related to epa_site to an internal epa_site dictionary."""
+        try:
+            internal = super().to_internal_value(data)
+            return internal
+        except ValidationError as exc:
+            raise exc
