+++ server/apps/trak/serializers/manifest_ser.py
@@ -8,7 +8,7 @@ from apps.trak.serializers.handler_ser import ManifestHandlerSerializer
 
 from ..models.manifest_models import AdditionalInfo
 from .base_ser import TrakBaseSerializer
-from .transporter_ser import TransporterSerializer
+from .handler_ser import TransporterSerializer
 from .waste_line_ser import WasteLineSerializer
 
 logger = logging.getLogger(__name__)
