+++ server/apps/trak/tests/serializers/test_manifest_serializers.py
@@ -1,5 +1,4 @@
-from apps.trak.models import Manifest, WasteLine
-from apps.trak.models.manifest_models import AdditionalInfo
+from apps.trak.models import AdditionalInfo, Manifest, WasteLine
 from apps.trak.serializers import ManifestSerializer
